+++ frontend/projects/repofluent-app/src/app/quality/performance-budget-page.component.ts
@@ -1,0 +1,12 @@
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
+import { PerformanceBudgetAdapter } from 'components';
+
+@Component({
+  selector: 'app-performance-budget-page',
+  templateUrl: './performance-budget-page.component.html',
+  styleUrl: './performance-budget-page.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class PerformanceBudgetPageComponent {
+  protected readonly performanceBudget = inject(PerformanceBudgetAdapter);
+}